mis en place multer, vérification role

+++ back/controllers/profile.js
@@ -1,6 +1,5 @@
 const dbConnection = require('../db/mysql.js');
-const cryptojs = require('crypto-js');
-const bcrypt = require('bcrypt');
+const fs = require('fs');
 
 exports.getprofile = (req, res) => {
 	dbConnection.query('SELECT * FROM users WHERE id = ?', req.params.id, (err, result) => {
@@ -8,7 +7,12 @@ exports.getprofile = (req, res) => {
 		if(result == 0){
             res.status(400).json({error: 'Invalid request !'});
         }; 
-		res.status(200).json(result[0]);
+        const dataUser = {
+            lastname: result[0].lastname,
+            firstname: result[0].firstname,
+            //email: result[0].email
+        }
+		res.status(200).json(dataUser);
 	});
 };
 
@@ -16,33 +20,54 @@ exports.deleteprofile = (req, res) => {
     dbConnection.query('SELECT * FROM users WHERE id = ?', req.params.id, (err, result) => {
         if(result == 0) res.status(400).json({error: 'Invalid request !'});
         else {
-            dbConnection.query('DELETE FROM users WHERE id = ?', req.params.id, (err, result) => {
-                if(err) throw err;
-                else res.status(200).json({message: 'User deleted !'});
-            });
+            const filename = result[0].imageUrl.split('/images/')[1];
+            if (filename !== "defaultPicture.jpg"){
+                fs.unlink(`images/${filename}`, () => {
+                    dbConnection.query('DELETE FROM users WHERE id = ?', req.params.id, (err, result) => {
+                        if(err) throw err;
+                        res.status(200).json({message: 'User deleted !'});
+                    });
+                });   
+            }
         }
     });	
 };
 
-//***************  Fonctionne pas *********************//
 exports.editProfile = (req, res) => {
-    if(req.auth.userId != req.params.id){
-        res.status(400).json({error: 'Invalid request !'});
-    } else {
-        //console.log(req.body)
-        bcrypt.hash(req.body.password, 10)
-            .then(hash => {
-                const editProfile = {
-                    name: req.body.name,
-                    lastname: req.body.lastname,
-                    email: cryptojs.HmacSHA256(req.body.email, `${process.env.PASSWORD_CRYPTOJS}`).toString(),
-                    password: hash
+    dbConnection.query('SELECT * FROM users WHERE id = ?', req.params.id, (err, result) => {
+        if(req.auth.userId != req.params.id || result == 0){
+            res.status(400).json({error: 'Invalid request !'});
+        }  
+        else {
+            const imageUploaded = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+            if (imageUploaded !== "http://localhost:3000/images/defaultPicture.jpg"){
+                dbConnection.query('UPDATE users SET imageUrl = ?', [imageUploaded, req.params.id], (err, result) => {
+                    if(err) throw err;
+                    res.status(200).json({message: 'User updated !'});
+                });
             }
-        dbConnection.query('UPDATE users SET ? WHERE id = ?', [editProfile, req.params.id], (err, result) => {
-            if(err) throw err;
-            else res.status(200).json({message: 'User updated !'});
-        });
+        }
     })
-        .catch(error => res.status(500).json({error}));
-    }
-};
+
+};   
+
+    
+    
+        
+
+        
+
+            // dbConnection.query('SELECT * FROM users WHERE id = ?', req.params.id, (err, result) => {
+            //     if (err) throw err;
+            //     if(result == 0){
+            //         res.status(400).json({error: 'Invalid request !'});
+            //     }; 
+            //     res.status(200).json(result[0]);
+            // });
+        
+    
+            
+       
+            
+        
+
